requirements.txt
Flask==3.0.3
openai==0.27.0
Werkzeug==3.0.6
python-dotenv==1.0.1